pyproject.toml
[project]
name = "chat-empresas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "boto3>=1.35.94",
    "duckdb>=1.1.3",
    "langchain>=0.3.14",
    "langchain-aws>=0.2.10",
    "langchain-community>=0.3.14",
    "langgraph>=0.2.61",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "polars>=1.19.0",
    "pyarrow>=18.1.0",
    "pyspark>=3.5.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.41.1",
    "tqdm>=4.67.1",
]
